requirements.txt
fastapi
uvicorn
pykrx
pandas
google-cloud-firestore
holidays